// src/api/generated/schemas/getBillingPricing200.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Elysia Documentation
 * Development documentation
 * OpenAPI spec version: 0.0.0
 */

import type { GetBillingPricing200CloudAi } from "./getBillingPricing200CloudAi";
import type { GetBillingPricing200FeaturesItem } from "./getBillingPricing200FeaturesItem";
import type { GetBillingPricing200LensPriceMicros } from "./getBillingPricing200LensPriceMicros";
import type { GetBillingPricing200LensPriceUsd } from "./getBillingPricing200LensPriceUsd";
import type { GetBillingPricing200Request } from "./getBillingPricing200Request";

export type GetBillingPricing200 = {
	currency: "USD";
	available: boolean;
	/** @nullable */
	lensPriceUsd: GetBillingPricing200LensPriceUsd;
	/** @nullable */
	lensPriceMicros: GetBillingPricing200LensPriceMicros;
	trialLenses: number;
	request: GetBillingPricing200Request;
	cloudAi: GetBillingPricing200CloudAi;
	features: GetBillingPricing200FeaturesItem[];
};